Make special ExitError.Error for Go <1.14
